+++ changelog.md
@@ -22,7 +22,9 @@ This version is not yet released. If you are reading this on the website, then t
 - Add `# No inline!` semantic comment, which prevents a function and its callers from being inlined
   - This enables stack traces on errors
 ### Interpreter
-- Add macro expansions as code actions to the language server
+- Language Server
+  - Add macro expansion as a code action
+  - Add completions for shadowable constants
 
 ## 0.9.5 - 2024-02-28
 ### Interpreter